# backend/app/routes/utente.js
const express = require("express");
const router = express.Router();
const { errorRes, successRes } = require("../response");
const GestoreProfilo = require("../gestori/GestoreProfilo");

const { db } = require("../db");
const User = require("../db/utente").User;
const ClasseUtente = require("../utils/ClasseUtente");
const Manager = require("../utenti/Manager");

// Route for getting user profile, only accessible if logged in
router.get("/profilo", async (req, res, next) => {
  const user = await User.findOne({ _id: req.user._id });

  successRes(res, "User profile", { user: user });
});

// Route for getting user profile, only accessible if logged in
router.post("/profilo", async (req, res, next) => {
  const user = await User.findOne({ _id: req.user._id });
  try {
    const response = await ClasseUtente.getClasseUtente(user.userType).creaAccount(
      req.body.nome,
      req.body.userType,
      req.body.email,
      req.body.password,
      ""
    );
    successRes(res, "User profile", { user: response });
  } catch (error) {
    errorRes(res, error, error.message, error.code);
  }

});

//Route for changing password
router.put("/password/:id?", async (req, res, next) => {
  try {
    const user = await User.findOne({ _id: req.user._id });

    let id =
      req.params.id != null && user.userType == "Manager"
        ? req.params.id
        : req.user._id;

    console.log("AAAAAAAA: " + user.userType);

    await ClasseUtente.getClasseUtente(user.userType).modificaPassword(
      id,
      req.body.vecchiaPassword,
      req.body.nuovaPassword
    );

    const response = await User.findOne({ _id: id });

    successRes(res, "Password cambiata con successo", { user: response });
  } catch (error) {
    errorRes(res, error, error.message, error.code);
  }
});

//Route for changing name
router.put("/nome/:id?", async (req, res, next) => {
  try {
    const user = await User.findOne({ _id: req.user._id });

    let id =
      req.params.id != null && user.userType == "Manager"
        ? req.params.id
        : req.user._id;

    await ClasseUtente.getClasseUtente(user.userType).modificaNome(
      id,
      req.body.nome
    );

    const response = await User.findOne({ _id: id });

    successRes(res, "Nome cambiato con successo", { user: response });
  } catch (error) {
    errorRes(res, error, "Impossibile cambiare il nome");
  }
});

module.exports = router;