# swagger.yaml
openapi: 3.0.0
info:
  version: "1.0"
  title: "BookEat"
  description: BookEat é una applicazione che permette agli utenti di
    visualizzare un menu ed ordinare le pietanze desiderate.
    Consente inoltre ai dipendenti del ristorante di gestire
    i vari ordini ed i posti a sedere comodamente da un tablet.
  license:
    name: MIT
    url: https://github.com/marcon21/bookeat/blob/main/LICENSE

servers:
  - url: http://localhost:3001/api/v1
    description: Local server
    
paths:
  /menu/{idPiatto}:
    get:
      summary: Ritorna tutti i piatti del menu
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  succes: 
                    type: boolean
                  message: 
                    type: string
                  data:
                    type: object
                    properties:              
                      categorie:
                        type: array
                      piatti:
                        type: array
                        items:
                          $ref: "#/components/schemas/Piatto"
              example:
                succes: true
                message: "OK"
                data: 
                  categorie: [{
                    "primaria": "Pizza"
                  }, { 
                    "primaria": "Primo",
                    "secondaria": "Mare" 
                  }]
                  piatti:
                    - _id: 1
                      nome: "Pizza Margherita"
                      prezzo: 500
                      categoria: {
                        "primaria": "Pizza"
                      }
                      disponibile: 1
                      descrizione: "Pizza con mozzarella e pomodoro"
                      allergeni: ["Glutine", "Lattosio"]
                      ingredientiModificabili:
                        ["Mozzarella", "Salamino", "Funghi"]
                    - _id: 2
                      nome: "Pizza Diavola"
                      categoria: {
                        "primaria": "Pizza"
                      }
                      prezzo: 600
                      disponibile: 1
                      descrizione: "Pizza con mozzarella, pomodoro e salame piccante"
                      allergeni: ["Glutine", "Lattosio"]
                      ingredientiModificabili:
                        ["Mozzarella", "Salamino", "Funghi"]

        "424":
          description: Errore nel recupero del menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                  success: false
                  error: "Errore nel recupero del menu"
    
    post:
      summary: Aggiunge un piatto al menu
      security:
        - jwtCookie: []
      requestBody:
        description: Il piatto da inserire nel menu
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Piatto"
            example:
              _id: 1
              nome: "Pizza Margherita"
              prezzo: 500
              categoria: {
                "primaria": "Pizza"
              }
              disponibile: 1
              descrizione: "Pizza con mozzarella e pomodoro"
              allergeni: ["Glutine", "Lattosio"]
              ingredientiModificabili:
                ["Mozzarella", "Salamino", "Funghi"]
      
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  succes: 
                    type: boolean
                  message: 
                    type: string
                  data: 
                    type: array
              example:
                succes: true
                message: "OK"
                data: []
          
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                  success: false
                  error: "Unauthorized - Accedi con un account autorizzato e riprova"   
                  
        "424":
          description: Creazione piatto fallita
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                  success: false
                  error: "Creazione piatto fallita"
    
    put:
      summary: Modifica un piatto identidificato da idPiatto
      security:
        - jwtCookie: []
      parameters:
        - in: path
          name: idPiatto
          schema:
            type: integer
          required: true
          description: Id del piatto da modificare
      requestBody:
        description: Il piatto da inserire nel menu
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Piatto"
            example:
              _id: 1
              nome: "Pizza Margherita"
              prezzo: 500
              categoria: {
                "primaria": "Pizza"
              }
              disponibile: 1
              descrizione: "Pizza con mozzarella e pomodoro"
              allergeni: ["Glutine", "Lattosio"]
              ingredientiModificabili:
                ["Mozzarella", "Salamino", "Funghi"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  succes: 
                    type: boolean
                  message: 
                    type: string
                  data: 
                    type: array
              example:
                succes: true
                message: "OK"
                data: []
                
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                  success: false
                  error: "Unauthorized - Accedi con un account autorizzato e riprova"
                  
        "424":
          description: Modifica piatto fallita
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                  success: false
                  error: "Modifica piatto fallita"
    
    delete:
      summary: Elimina un piatto identificato da idPiatto
      security:
        - jwtCookie: []
      parameters:
        - in: path
          name: idPiatto
          schema:
            type: integer
          required: true
          description: Id del piatto da eliminare
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  succes: 
                    type: boolean
                  message: 
                    type: string
                  data: 
                    type: array
              example:
                succes: true
                message: "OK"
                data: []
                
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                  success: false
                  error: "Unauthorized - Accedi con un account autorizzato e riprova"
                  
        "424":
          description: Eliminazione piatto fallita
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                  success: false
                  error: "Eliminazione piatto fallita"
  
  /conto/apriConto:
    post:
      summary: Apre un conto
      security:
        - jwtCookie: []
      requestBody:
        description: Il numero di coperti 
        content:
          application/json:
            schema:
              type: integer
            example:
              nCoperti: 4
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  succes: 
                    type: boolean
                  message: 
                    type: string
                  data: 
                    type: string
                    description: Id del conto aperto
              example:
                succes: true
                message: "OK"
                data: "6467b502aa1394c96fe45299"
          
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                  success: false
                  error: "Unauthorized - Accedi con un account autorizzato e riprova"   
                  
        "424":
          description: Apertura conto fallita
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                  success: false
                  error: "Apertura conto fallita"
      
  /conto/invioOrdine:
    post:
      summary: Invia un ordine
      security:
        - jwtCookie: []
      requestBody:
        description: L'ordine da inviare
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  idPiatto:
                    type: string
                  ingredientiScelti:
                    type: array
                  note:
                    type: string
                  priorita:
                    type: integer
              
            example: 
              - idPiatto: "6467b502aa1394c96fe45299"
                ingredientiScelti: []
                note: ""
                prioritá: 0
              
              - idPiatto: "6467b502aa1394c96fe45299"
                ingredientiScelti: []
                note: ""
                prioritá: 0
              
              - idPiatto: "6467bhb1231394c96fe45299"
                ingredientiScelti: []
                note: ""
                prioritá: 2
      
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  succes: 
                    type: boolean
                  message: 
                    type: string
                  data: 
                    type: array
              example:
                succes: true
                message: "OK"
                data: []
          
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                  success: false
                  error: "Unauthorized - Accedi con un account autorizzato e riprova"   
                  
        "424":
          description: Creazione piatto fallita
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                  success: false
                  error: "Creazione piatto fallita"
  
  /auth/login:
    post:
      summary: Path per loggarsi e rivecere il token jwt
      requestBody:
        description: Body con campo *email* e *password*
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Utente"
            example:
              email: "example@email.com"
              password: "password"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                      type: boolean
                  message:
                      type: string
                  data:
                    type: object
                    properties:
                      token: 
                        type: string
                      userType:
                        type: string
              example:
                success: true
                message: Account creato correttamente
                data:
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7Il9pZCI6IjY0NjAkjnkaszJlM2UxYmYxNGIyYzEzZSIsImVtYWlsIjoiZGFuaWVsLm1hcmNvbkBzdHVkZW50aS51bml0bi5pdCJ9LCJpYXQiOjE2ODQzMzExNTJ9.fjfOnRwwunj9BuzhFiN_GpNCJajfAKnYqw3yFTVZS8U"
                  userType: "UtenteLoggato"
        "401":
            description: Bad credentials
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrResponse"
                example:
                  success: false
                  error: "Errore, credenziali errate"
  
  /auth/signup:
    post:
      summary: Path per registrarsi e rivecere il token jwt
      requestBody:
        description: Body con campo *email* e *passoword*
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Utente"
            example:
              email: "example@email.com"
              password: "password"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                      type: boolean
                  message:
                      type: string
                  data:
                    type: object
                    properties:
                      token: 
                        type: string
              example:
                success: true
                message: Account creato correttamente
                data:
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7Il9pZCI6IjY0NjAkjnkaszJlM2UxYmYxNGIyYzEzZSIsImVtYWlsIjoiZGFuaWVsLm1hcmNvbkBzdHVkZW50aS51bml0bi5pdCJ9LCJpYXQiOjE2ODQzMzExNTJ9.fjfOnRwwunj9BuzhFiN_GpNCJajfAKnYqw3yFTVZS8U"
                  userType: "UtenteLoggato"
        "401":
            description: Bad credentials
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrResponse"
                example:
                  success: false
                  error: "Errore, credenziali errate"
  
  /utente/profilo:
    get:
      summary: Ricevi informazioni sul profilo dell'utente
      security:
        - jwtCookie: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                      type: boolean
                  message:
                      type: string
                  data:
                    $ref: "#/components/schemas/UtenteResponse"
        "401":
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                success: false
                error: "Utente non autenticato"
                /utente/profilo:
    post:
      summary: Creazione di un nuovo utente, accessibile sola al manager
      security:
        - jwtCookie: []
      requestBody:
        description: Nuovo Utente
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                email:
                  type: string
                password:
                  type: string
                userTyp:
                  type: string
      responses:
        "200":
          description: Response with the user created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                      type: boolean
                  message:
                      type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: "#/components/schemas/UtenteResponse"
        "401":
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                success: false
                error: "Utente non autenticato"

  /utente/password/{?idUtente}:
    put:
      summary: Modifica password dell'utente
      security:
        - jwtCookie: []
      requestBody:
        description: Nuova password dell'utente
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RichiestaCambioPassword"
            example:
              vecchiaPassword: "password1234"
              nuovaPassword: "newpassword1234"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  succes: 
                    type: boolean
                  message: 
                    type: string
                  data:
                    $ref: "#/components/schemas/UtenteResponse"
              example:
                success: true
                message: "Password modificata con successo"

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                success: false
                error: "Richiesta non valida"
        
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                success: false
                error: "Unauthorized - Accedi"
                
  /utente/nome/{?idUtente}:
    put:
      summary: Modifica il nome dell'utente
      security:
        - jwtCookie: []
      requestBody:
        description: Nuovo nome dell'utente
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RichiestaCambioNome"
            example:
              nome: "Glauco"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  succes: 
                    type: boolean
                  message: 
                    type: string
                  data:
                    $ref: "#/components/schemas/UtenteResponse"
              example:
                success: true
                message: "Nome cambiato con successo"

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                success: false
                error: "Richiesta non valida"
        
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrResponse"
              example:
                success: false
                error: "Unauthorized - Accedi"

components:
  schemas:
    Piatto:
      type: object
      properties:
        _id:
          type: string
          description: ID del piatto
        nome:
          type: string
          description: Nome del piatto
        prezzo:
          type: integer
          description: Prezzo del piatto
        categoria:
          type: object
          properties:
            primaria:
              type: string
            secondaria:
              type: string
          description: La categoria del piatto
        disponibile:
          type: number
          description: Disponibilità del piatto
        descrizione:
          type: string
          description: Descrizione del piatto
        allergeni:
          type: array
          description: Allergeni del piatto
        ingredientiModificabili:
          type: array
          description: Ingredienti modificabili del piatto
      example:
        _id: 1
        nome: "Pizza Margherita"
        prezzo: 500
        categoria: "Pizze"
        disponibile: 1
        descrizione: "Pizza con mozzarella e pomodoro"
        allergeni: ["Glutine", "Lattosio"]
        ingredientiModificabili: ["Mozzarella", "Salamino", "Funghi"]
    
    Portata: 
      type: object
      properties:
        _id:
          type: string
          description: Id della portata
        nome:
          type: string
          description: Il nome della portata
        prezzo:
          type: integer
          description: il prezzo della portata
        ingredientiScelti:
          type: array
          description: Ingredienti opzionali selezionati dall'utente
        note:
          type: string
          description: Note addizionali scritte dall'utente
        status:
          type: integer
          description: Lo status della portata
        prioritá:
          type: integer
          description: La prioritá della portata
    
    Utente:
      type: object
      properties:
        email:
          type: string
          description: Campo email
        password:
          type: string
          description: Campo password
      example:
        email: "example@email.com"
        password: "password"
        
    ErrResponse:
      type: object
      properties:
        success:
            type: boolean
        error:
            type: string

    RichiestaCambioPassword:
      type: object
      properties:
        vecchiaPassword:
          type: string
          description: Vecchia password dell'utente
        nuovaPassword:
          type: string
          description: Nuova password dell'utente
      required:
        - vecchiaPassword
        - nuovaPassword
        
    RichiestaCambioNome:
      type: object
      properties:
        nome:
          type: string
          description: Nuovo nome dell'utente
      required:
        - nome
        
    UtenteResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            _id:
              type: string
            email:
              type: string
            password:
              type: string
            nome:
              type: string
            userType:
              type: string
            googleId:
              type: string
            
  securitySchemes:
    jwtCookie:         
      type: apiKey
      in: cookie
      name: jwt